processing-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiokafka==0.11.0
httpx==0.27.0
python-decouple==3.8
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio==4.4.0
coverage==7.5.3
